+++ src/roll.cpp
@@ -271,6 +271,24 @@ arma::uvec any_na_xy(const NumericMatrix& x, const NumericMatrix& y) {
   
 }
 
+arma::ivec stl_sort_index(arma::vec& x) {
+  
+  int n_rows_x = x.size();
+  arma::ivec y(n_rows_x);
+  std::iota(y.begin(), y.end(), 0);
+  
+  auto comparator = [&x](int a, int b) {
+    if (std::isnan(x[a])) return false;
+    if (std::isnan(x[b])) return true;
+    return x[a] < x[b];
+  };
+  
+  std::sort(y.begin(), y.end(), comparator);
+  
+  return y;
+  
+}
+
 // [[Rcpp::export(.roll_any)]]
 LogicalMatrix roll_any(const LogicalMatrix& x, const int& width,
                        const int& min_obs, const bool& complete_obs,