+++ NEWS.md
@@ -1,6 +1,6 @@
 ## News for package 'roll'
 
-### Changes in roll version 1.1.2 (YYYY-MM-DD)
+### Changes in roll version 1.1.2 (2019-MM-DD)
 
 * New `roll_median`, `roll_min`, `roll_max`, `roll_any`, and `roll_all` functions for computing rolling medians, minimums, maximums, any, and all, respectively, of time-series data ([#4](https://github.com/jjf234/roll/issues/4), [#13](https://github.com/jjf234/roll/issues/13), [#14](https://github.com/jjf234/roll/issues/14))
     * Note: `roll_median`, `roll_min`, and `roll_max` functions are not calculated using online algorithms